Some code and include refractoring and implemented a working VMM :) .

+++ Includes/LibC/string.h
@@ -4,8 +4,8 @@
 #include <stddef.h>
 #include <stdint.h>
 
-void* memset(void* ptr, int value, size_t count);
-void* memsetw(void* ptr, int value, size_t count);
+void* memset(void* ptr, uint8_t value, size_t count);
+void* memsetw(void* ptr, uint16_t value, size_t count);
 void* memsetq(void* ptr, uint64_t value, size_t count);
 
 void* memcpy(void* dest, void* src, size_t count);
